done adding aux_input to core
